+++ tests/unit_tests/test_parser.py
@@ -20,9 +20,10 @@ from typing import Callable
 from unittest import mock
 
 import pytest
-from tests.unit_tests.util import assert_ellipsis_match
+from util import assert_ellipsis_match
 
 from croud import __version__
+from croud.__main__ import get_parser
 from croud.parser import CroudCliArgumentParser, CroudCliHelpFormatter, create_parser
 
 
@@ -229,10 +230,38 @@ Optional Arguments:
             "local",
             "--output-fmt",
             "json",
-            "--sudo",
         ]
         args = parser.parse_args(argv)
         assert args.env == "local"
         assert args.region == "eastus.azure"
         assert args.output_fmt == "json"
+
+    def test_default_sudo_argument(self):
+        tree = {"help": "help text", "commands": {"cmd": {"resolver": noop}}}
+        parser = create_parser(tree)
+
+        argv = ["cmd", "--sudo"]
+        args = parser.parse_args(argv)
         assert args.sudo is True
+
+    def test_config_set_help_text(self, capsys):
+        parser = get_parser()
+        argv = ["config", "set", "-h"]
+        with pytest.raises(SystemExit):
+            parser.parse_args(argv)
+        out, _ = capsys.readouterr()
+        assert "--sudo" not in out
+        assert "--region" in out
+        assert "--env" in out
+        assert "--output-fmt" in out
+
+    def test_config_get_help_text(self, capsys):
+        parser = get_parser()
+        argv = ["config", "get", "-h"]
+        with pytest.raises(SystemExit):
+            parser.parse_args(argv)
+        out, _ = capsys.readouterr()
+        assert "--sudo" not in out
+        assert "--region" in out
+        assert "--env" in out
+        assert "--output-fmt" in out